fix(runtime): ask the kernel whether a process is alive, not /proc
`ServerHealthChecker::is_process_alive` was `#[cfg(unix)]` and implemented as
`Path::new(&format!("/proc/{pid}")).exists()`. macOS is `cfg(unix)` and has no
`/proc`, so on macOS every PID would have read as dead: `check_process` returns
`ProcessDied`, `check_combined` short-circuits on that and never attempts the
HTTP check, and `ServerHealthStatus::is_failed` reports the server failed.
`aarch64-apple-darwin` is a release target.

The fix already existed sixty lines away. `pidfile::pid_exists` asks the kernel
with a null signal and handles ESRCH/EPERM, and had no production caller at all —
only a test. The unix arm now delegates to it.

Scope, stated precisely: the bug was **latent, not observed**. The only
production caller of `ServerHealthMonitor` builds its `ProcessHandle` with
`pid: None` (`gglib-app-services/src/servers.rs:266`), and `check_process`
returns `Healthy` on that branch without consulting `is_process_alive` at all.
The one other source of a real PID, `ProcessManager::list_running`, reaches only
the server list. Supplying a PID — which the public API permits, and which
`list_running`'s own doc already advertises as a consumer — is all it would have
taken.

The `cfg(not(unix))` arm deliberately still returns `true` rather than
delegating. `pid_exists`'s non-unix arm returns `false` ("not implemented"), so
delegating there would report every Windows server that supplied a PID as
`ProcessDied` — transplanting to `x86_64-pc-windows-msvc`, also a release
target, the exact bug this removes from macOS. That body is byte-identical to
what ships today; only its comment is new.

Also removes `check_process_health` and `update_health_batch`: two copy-pasted
sysinfo variants of the same question, with zero callers anywhere in the tree
outside the blanket re-export they sat behind. `sysinfo` stays — `system/gpu.rs`
and `pidfile/verify.rs` still use it.

Pinned by a test watched to fail first: reverting the delegation makes
`a_live_process_reports_healthy_rather_than_dead` fail and nothing else. The
existing `test_process_check_with_invalid_pid` could not have caught this — with
`/proc` absent every PID reads as dead, so asserting that an invalid PID is dead
passed for the wrong reason. The new guard is vacuous on CI for the mirror-image
reason, and that is recorded in the test rather than left to be rediscovered:
only `ubuntu-latest` runs this crate's tests, and `/proc` exists there. It fails
only on a macOS machine.

The test module moves to a `#[path]` sibling following #860 — health_monitor.rs
had ten lines of headroom under the 300 LOC budget and this needed more.

Ratchet: health_monitor.rs 280 -> 204, process/health.rs 163 -> 125,
health_monitor_tests.rs 119 recorded. None was on the baseline and none is now.
